requirements.txt
django
pyopenssl
psycopg2-binary
djangorestframework
